requirements.txt
asgiref==3.2.3
boto==2.49.0
boto3==1.12.13
botocore==1.15.13
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
django-mysql==3.3.0
docutils==0.15.2
docx==0.2.4
gensim==3.8.1
idna==2.9
jmespath==0.9.5
lxml==4.5.0
nltk==3.4.5
numpy==1.18.1
pdfminer==20191125
Pillow==7.0.0
pycryptodome==3.9.7
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.9.0
sqlparse==0.3.1
urllib3==1.25.8
XlsxWriter==1.2.8
